Update getInfoFromDs to support connecting to CA in a remote forest (#30)
Removed the conditional statement on the 3rd line of the function as it seemed unneeded. Updated retrieval of forest name to query the domain of the CA and rootDomainNamingContext via the RootDSA using LDAP. Before this was assuming the forest to the be same of the user running the function. This would cause attempting to connect to a CA in a resource forest to fail if using the account in the forest for accounts. Updated displayName, enrollemtn servers to be retrieved via LDAP as it is published in AD and in my test worked better then using _certConfig.

Example of getting a CA in a remote forest in Powershell.
$ca = [PKI.CertificateServices.CertificateAuthority]::new('COMPUTER_NAME', 'CA_DISPLAYNAME');
$ca | Get-IssuedRequest